Refactor API structure by removing unused auth module, conditionally including test router based on environment variable, and updating test endpoints for Firebase connection and token verification
